Reduce code duplication, other sonarlint issues
