+++ FirstVoicesCoreTests/src/main/resources/OSGI-INF/mocks/mocks-operations.xml
@@ -18,21 +18,12 @@
   </extension>
 
   <extension point="operations" target="org.nuxeo.ecm.core.operation.OperationServiceComponent">
-    <operation class="ca.firstvoices.tests.mocks.operations.GenerateUsersForDialect"/>
+    <operation class="ca.firstvoices.tests.mocks.operations.GenerateUsers"/>
   </extension>
 
   <extension point="operations" target="org.nuxeo.ecm.core.operation.OperationServiceComponent">
-    <operation class="ca.firstvoices.tests.mocks.operations.GenerateUsersForDialects"/>
+    <operation class="ca.firstvoices.tests.mocks.operations.RemoveUsers"/>
   </extension>
-
-  <extension point="operations" target="org.nuxeo.ecm.core.operation.OperationServiceComponent">
-    <operation class="ca.firstvoices.tests.mocks.operations.RemoveUsersForDialect"/>
-  </extension>
-
-  <extension point="operations" target="org.nuxeo.ecm.core.operation.OperationServiceComponent">
-    <operation class="ca.firstvoices.tests.mocks.operations.RemoveUsersForDialects"/>
-  </extension>
-
 
   <!-- Block Operation from being visible to non super admins -->
   <extension target="org.nuxeo.ecm.automation.server.AutomationServer" point="bindings">
@@ -48,16 +39,10 @@
     <binding name="Mocks.RemoveDialects">
       <administrator>true</administrator>
     </binding>
-    <binding name="Mocks.GenerateUsersForDialect">
+    <binding name="Mocks.GenerateUsers">
       <administrator>true</administrator>
     </binding>
-    <binding name="Mocks.GenerateUsersForDialects">
-      <administrator>true</administrator>
-    </binding>
-    <binding name="Mocks.RemoveUsersForDialect">
-      <administrator>true</administrator>
-    </binding>
-    <binding name="Mocks.RemoveUsersForDialects">
+    <binding name="Mocks.RemoveUsers">
       <administrator>true</administrator>
     </binding>
   </extension>
