+++ FirstVoicesCoreTests/src/test/java/ca/firstvoices/operations/GenerateUsersTest.java
@@ -2,7 +2,7 @@ package ca.firstvoices.operations;
 
 import ca.firstvoices.testUtil.AbstractFirstVoicesCoreTestsTest;
 import ca.firstvoices.tests.mocks.operations.GenerateDialect;
-import ca.firstvoices.tests.mocks.operations.GenerateUsersForDialects;
+import ca.firstvoices.tests.mocks.operations.GenerateUsers;
 import java.util.HashMap;
 import java.util.List;
 import org.junit.Assert;
@@ -11,7 +11,38 @@ import org.nuxeo.ecm.automation.OperationContext;
 import org.nuxeo.ecm.automation.OperationException;
 import org.nuxeo.ecm.core.api.DocumentModel;
 
-public class GenerateUsersForDialectsTest extends AbstractFirstVoicesCoreTestsTest {
+public class GenerateUsersTest extends AbstractFirstVoicesCoreTestsTest {
+
+  @Test
+  public void generateUsersForDialect() throws OperationException {
+    params = new HashMap<>();
+    params.put("randomize", "false");
+    params.put("maxEntries", "50");
+    params.put("dialectName", "Xx_Dialect_xX");
+
+    OperationContext ctx = new OperationContext(session);
+
+    DocumentModel dialect = (DocumentModel) automationService.run(ctx, GenerateDialect.ID, params);
+
+    createNewGroup(dialect.getName().toLowerCase() + "_members",
+        dialect.getName() + " Members");
+    createNewGroup(dialect.getName().toLowerCase() + "_recorders",
+        dialect.getName() + " Recorders");
+    createNewGroup(dialect.getName().toLowerCase() + "_recorders_with_approval",
+        dialect.getName() + " Recorders With Approval");
+    createNewGroup(dialect.getName().toLowerCase() + "_language_administrators",
+        dialect.getName() + " Language Administrators");
+
+    params = new HashMap<>();
+    params.put("dialectName", "Xx_Dialect_xX");
+    automationService.run(ctx, GenerateUsers.ID, params);
+
+    List<String> existingUsers = userManager.getUserIds();
+    Assert.assertTrue(existingUsers.contains(dialect.getName() + "_members"));
+    Assert.assertTrue(existingUsers.contains(dialect.getName() + "_recorders"));
+    Assert.assertTrue(existingUsers.contains(dialect.getName() + "_recorders_with_approval"));
+    Assert.assertTrue(existingUsers.contains(dialect.getName() + "_language_administrators"));
+  }
 
   @Test
   public void generateUsersForDialects() throws OperationException {
@@ -49,7 +80,7 @@ public class GenerateUsersForDialectsTest extends AbstractFirstVoicesCoreTestsTe
     createNewGroup(dialectB.getName().toLowerCase() + "_language_administrators",
         dialectB.getName() + " Language Administrators");
 
-    automationService.run(ctx, GenerateUsersForDialects.ID);
+    automationService.run(ctx, GenerateUsers.ID);
 
     List<String> existingUsers = userManager.getUserIds();
     Assert.assertTrue(existingUsers.contains(dialectA.getName() + "_members"));
@@ -61,8 +92,5 @@ public class GenerateUsersForDialectsTest extends AbstractFirstVoicesCoreTestsTe
     Assert.assertTrue(existingUsers.contains(dialectB.getName() + "_recorders"));
     Assert.assertTrue(existingUsers.contains(dialectB.getName() + "_recorders_with_approval"));
     Assert.assertTrue(existingUsers.contains(dialectB.getName() + "_language_administrators"));
-
-
   }
-
 }
