+++ FirstVoicesCoreTests/src/test/java/ca/firstvoices/operations/RemoveUsersTest.java
@@ -2,8 +2,8 @@ package ca.firstvoices.operations;
 
 import ca.firstvoices.testUtil.AbstractFirstVoicesCoreTestsTest;
 import ca.firstvoices.tests.mocks.operations.GenerateDialect;
-import ca.firstvoices.tests.mocks.operations.GenerateUsersForDialects;
-import ca.firstvoices.tests.mocks.operations.RemoveUsersForDialects;
+import ca.firstvoices.tests.mocks.operations.GenerateUsers;
+import ca.firstvoices.tests.mocks.operations.RemoveUsers;
 import java.util.HashMap;
 import java.util.List;
 import org.junit.Assert;
@@ -12,7 +12,51 @@ import org.nuxeo.ecm.automation.OperationContext;
 import org.nuxeo.ecm.automation.OperationException;
 import org.nuxeo.ecm.core.api.DocumentModel;
 
-public class RemoveUsersForDialectsTest extends AbstractFirstVoicesCoreTestsTest {
+public class RemoveUsersTest extends AbstractFirstVoicesCoreTestsTest {
+
+  @Test
+  public void removeUsersForDialect() throws OperationException {
+    params = new HashMap<>();
+    params.put("randomize", "false");
+    params.put("maxEntries", "50");
+    params.put("dialectName", "Xx_Dialect_xX");
+
+    OperationContext ctx = new OperationContext(session);
+
+    DocumentModel dialect = (DocumentModel) automationService.run(ctx, GenerateDialect.ID, params);
+
+    createNewGroup(dialect.getName().toLowerCase() + "_members",
+        dialect.getName() + " Members");
+    createNewGroup(dialect.getName().toLowerCase() + "_recorders",
+        dialect.getName() + " Recorders");
+    createNewGroup(dialect.getName().toLowerCase() + "_recorders_with_approval",
+        dialect.getName() + " Recorders With Approval");
+    createNewGroup(dialect.getName().toLowerCase() + "_language_administrators",
+        dialect.getName() + " Language Administrators");
+
+    params = new HashMap<>();
+    params.put("dialectName", "Xx_Dialect_xX");
+    automationService.run(ctx, GenerateUsers.ID, params);
+
+    List<String> existingUsers = userManager.getUserIds();
+    Assert.assertTrue(existingUsers.contains(dialect.getName() + "_members"));
+    Assert.assertTrue(existingUsers.contains(dialect.getName() + "_recorders"));
+    Assert.assertTrue(existingUsers.contains(dialect.getName() + "_recorders_with_approval"));
+    Assert.assertTrue(existingUsers.contains(dialect.getName() + "_language_administrators"));
+
+    automationService.run(ctx, RemoveUsers.ID, params);
+    Assert.assertTrue(
+        userManager.getUsersInGroup(dialect.getName().toLowerCase() + "_members").isEmpty());
+    Assert.assertTrue(
+        userManager.getUsersInGroup(dialect.getName().toLowerCase() + "_recorders").isEmpty());
+    Assert.assertTrue(
+        userManager.getUsersInGroup(dialect.getName().toLowerCase() + "_recorders_with_approval")
+            .isEmpty());
+    Assert.assertTrue(
+        userManager.getUsersInGroup(dialect.getName().toLowerCase() + "_language_administrators")
+            .isEmpty());
+
+  }
 
   @Test
   public void removeUsersForDialects() throws OperationException {
@@ -50,7 +94,7 @@ public class RemoveUsersForDialectsTest extends AbstractFirstVoicesCoreTestsTest
     createNewGroup(dialectB.getName().toLowerCase() + "_language_administrators",
         dialectB.getName() + " Language Administrators");
 
-    automationService.run(ctx, GenerateUsersForDialects.ID);
+    automationService.run(ctx, RemoveUsers.ID);
 
     List<String> existingUsers = userManager.getUserIds();
     Assert.assertTrue(existingUsers.contains(dialectA.getName() + "_members"));
@@ -63,7 +107,7 @@ public class RemoveUsersForDialectsTest extends AbstractFirstVoicesCoreTestsTest
     Assert.assertTrue(existingUsers.contains(dialectB.getName() + "_recorders_with_approval"));
     Assert.assertTrue(existingUsers.contains(dialectB.getName() + "_language_administrators"));
 
-    automationService.run(ctx, RemoveUsersForDialects.ID);
+    automationService.run(ctx, RemoveUsers.ID);
     Assert.assertTrue(
         userManager.getUsersInGroup(dialectA.getName().toLowerCase() + "_members").isEmpty());
     Assert.assertTrue(
